requirements.txt
tqdm
pytorch
pytorch-cuda
pytorch-metric-learning
pytorch-mutex
numpy
pandas
scikit-learn